% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canada_cd_pop}
\alias{canada_cd_pop}
\title{Canadian Census District Populations}
\format{
A \code{data.table} with 293 rows and 3 variables:
\describe{
  \item{unweighted_pop}{Population from the 2016 Canadian census.}
  \item{state_fips}{Province SGC for the county}
  \item{county_fips}{Census division code}
}
}
\source{
\url{https://www12.statcan.gc.ca/census-recensement/2016/dp-pd/hlt-fst/pd-pl/Tables/CompFile.cfm?Lang=Eng&T=701&OFT=FULLCSV}
}
\usage{
canada_cd_pop
}
\description{
Population by census district (with \code{state_fips} and \code{county_fips} identifiers to link with other data sets in the package - sorry for the naming, Canadians). The "unweighted" in the variable name \code{unweighted_pop} doesn't refer to anything specific in the Canadian census, but is so you can easily \code{rbind} this with \code{county_pop}.
}
\details{
This comes from the Canadian census directly instead of SafeGraph.
}
\keyword{datasets}
