% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canada_cd_types}
\alias{canada_cd_types}
\title{Additional census division information for Canada}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 293 rows and 5 columns.
}
\usage{
canada_cd_types
}
\description{
A dataset that can be merged with \code{fips_to_names} with information on French names for locations as well as the type of census division it is.
}
\details{
\describe{
  \item{state_fips}{Canadian SGC code}
  \item{county_fips}{Canadian Census division}
  \item{countyname_french}{Census division name in French}
  \item{cd_type}{Census district type in English}
  \item{cd_type_french}{Census district type in French}
}
}
\keyword{datasets}
