% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_pop}
\alias{county_pop}
\title{Unweighted Population by County}
\format{
A \code{data.table} with 3220 rows and 3 variables:
\describe{
  \item{unweighted_pop}{Population from the 2016 American Community Survey (the "unweighted" part is outdated but kept for consistency with old code).}
  \item{state_fips}{State FIPS code for the census block group}
  \item{county_fips}{County FIPS code for the census block group}
}
}
\source{
\url{https://docs.safegraph.com/docs/open-census-data}
}
\usage{
county_pop
}
\description{
A dataset containing the unweighted population by county from the Open Census file. Merge with \code{data(fips_to_names)} to name the states and counties.
}
\keyword{datasets}
