% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distancing}
\alias{distancing}
\title{Example Stay-at-Home Data}
\format{
A \code{data.table} with 45158 rows and 7 variables:
\describe{
  \item{date}{The date}
  \item{state_fips,county_fips}{The state and county identifiers}
  \item{device_count}{The total number of devices observed}
  \item{completely_home_device_count}{The total number of devices observed that did not leave their home location on this day}
  \item{part_time_work_behavior_devices,full_time_work_behavior_devices}{The total number of devices observed that appear to be engaging in part-time or full-time work behavior (experimental)}
}
}
\source{
SafeGraph
}
\usage{
distancing
}
\description{
Distancing data from June 1 to June 14, aggregated to the county level for the vignette. See [SafeGraph Docs](https://docs.safegraph.com/docs) for full documentation.
}
\keyword{datasets}
