% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_json.R
\name{expand_integer_json}
\alias{expand_integer_json}
\title{Expand and collapse an integer JSON column}
\usage{
expand_integer_json(
  dt,
  expand,
  index = "index",
  by = NULL,
  fast = TRUE,
  fun = sum,
  na.rm = TRUE,
  set_key = TRUE
)
}
\arguments{
\item{dt}{data.table object (or something that can be coerced to data.table)}

\item{expand}{String indicating the JSON column to be expanded.}

\item{index}{String indicating the name of the new index column}

\item{by}{Character vector indicating the variables to group by after expanding. Set to \code{NULL} to aggregate across all initial rows, or set to \code{FALSE} to not aggregate at all (this will also add an \code{initial_rowno} column showing the original row number).}

\item{fast}{Assumes that all the JSON vectors are of equal length, and adds the values together rather than using whatever is in \code{fun}.}

\item{fun}{Function that takes a vector and returns a single value to use when collapsing to the \code{by} level. Requires \code{fast = FALSE}.}

\item{na.rm}{Ignore missing values of \code{expand}}

\item{set_key}{Set the key of \code{dt} to \code{by}. Set to \code{FALSE} if you have already set the key or want it returned without key.}
}
\description{
This function accepts a \code{data.table} along with a set of grouping variables and a character-format integer-style JSON column (i.e. starts with square brackets, not curly).
}
\details{
It expands that JSON column into long format, with one row per observation per value of the JSON column, and then collapses everything according to the set of grouping variables.
}
\examples{

# Example data
patterns <- data.table::data.table(state_fips = c(1,1,2,2),
                                     int_origin = c('[2,3]',
                                                    '[3,4]',
                                                    '[4,5]',
                                                    '[5,6]'))

expand_integer_json(patterns, 'int_origin', by = 'state_fips')[]

}
