% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{expand_open_hours}
\alias{expand_open_hours}
\title{Expands the open_hours variable in the Core file}
\usage{
expand_open_hours(
  dt,
  format = c("wide", "long", "long-expand", "long_expand"),
  open_hours = "open_hours",
  colnames = NULL,
  drop_missing = FALSE,
  convert_hour = TRUE,
  days = c("Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday",
    "Sunday")
)
}
\arguments{
\item{dt}{A \code{data.table} containing the \code{open_hours} column (or an object that can be coerced to a \code{data.table}).}

\item{format}{Can be \code{'wide'} (seven new \code{list}-columns, one for each day), \code{'long'} (turn each row into seven rows, then two new columns: one for day-of-week and one \code{list}-column with opening/closing times), or \code{'long-expand'}/\code{'long_expand'} (\code{'long'} but then also break the single list-column out into a set of numeric start/end columns). Note that for \code{'long-expand'}, many locations have more than one set of open/close hours per day, so there will be more than one open/close column each.}

\item{open_hours}{A character variable with the name of the \code{open_hours} column.}

\item{colnames}{For \code{format = 'wide'}, the name stub for the column names, by default \code{'open_hours'} to get \code{'open_hoursSunday'}, \code{'open_hoursMonday'}, etc.. For \code{format='long'}, a two-element vector (by default \code{c('weekday','open_hours')}) with the name of the column indicating the day, and the \code{list}-column with the open hours information in it. For \code{format = 'long-expand'}, a three-element vector with the weekday column, the name stub for "opening hour" and the name stub for "closing hour" (with numbers 1, 2, 3, etc. appended afterwards), by default \code{c('weekday','opens','closes')}.}

\item{drop_missing}{Drop any rows with a missing \code{open_hours} observation.}

\item{convert_hour}{Convert hour strings like \code{'15:30'} to numbers like \code{15.5}. This does slow down the function.}

\item{days}{A character vector of the days to keep. Cutting down here can save some time/memory especially if you are not going \code{format = 'wide'}.}
}
\description{
This function takes the \code{open_hours} variable in an already-read Core file (stored as a \code{data.table}) and expands it to seven \code{list}-type columns, where the elements of the list in each row are a set of vectors for opening/closing times, in military time format (1:30PM = 13.5). So an observation of \code{c(8,10,12,14)} would be a business that opens at 8, closes at 10, opens again at noon, and closes again at 2PM on that day. Options are available to produce long instead of wide expansions as well.
}
\details{
Returns the same \code{data.table} but with the new columns/rows added. May change the order of the data.
}
