% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_patterns.R
\name{find_date}
\alias{find_date}
\title{Find the date in a SafeGraph AWS-formatted filepath}
\usage{
find_date(s)
}
\arguments{
\item{s}{The filepath to look for a date in.}
}
\description{
Given a filepath \code{s}, this function will look for the last correctly-parsed date in that string. Given how SafeGraph AWS file structures are, this will give you the date of those files, for example \code{patterns_backfill/2020/12/14/21/2018/01/01} will give you "2018/01/01".
}
\details{
This function returns a string, not a date. You may want to send it to \code{as.Date()} or \code{lubridate::ymd}.

For backfill data, the date returned will generally be the \code{start_date} for the files. However, for new data, you will want to do \code{as.Date(find_date(s)) - lubridate::days(9)} to get the \code{start_date}.
}
\examples{

start_date <- find_date('patterns_backfill/2020/12/14/21/2018/01/01') \%>\% as.Date()

}
