% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fips_from_cbg}
\alias{fips_from_cbg}
\title{Pull state and county FIPS from CBG code}
\usage{
fips_from_cbg(cbg, return = "both")
}
\arguments{
\item{cbg}{CBG code, in numeric or string form.}

\item{return}{Set to 'state' to get back only state FIPS, 'county' for only county, or 'both' for a list of both (state then county).}
}
\description{
This function takes a CBG code (as numeric or string) and returns the state and county FIPS codes associated with it.
}
\details{
Why a list and not a vector? For \code{data.table} usage.
}
