% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fips_to_names}
\alias{fips_to_names}
\title{State and county names by FIPS codes}
\format{
A \code{data.table} with 3142 rows and 4 variables:
\describe{
  \item{state_fips}{State FIPS code}
  \item{county_fips}{County FIPS code}
  \item{statename}{The full name of the state}
  \item{countyname}{The full name of the county, including "County"}
}
}
\source{
\url{US Census}
}
\usage{
fips_to_names
}
\description{
A dataset that linkes state and county FIPS codes (as numeric values) to the names of those states and counties
}
\keyword{datasets}
