% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_poi_naics.R
\name{link_poi_naics}
\alias{link_poi_naics}
\title{Use a Core Places file to Create a POI-NAICS crosswalk}
\usage{
link_poi_naics(
  filename,
  dir = ".",
  exdir = dir,
  cleanup = FALSE,
  silent = FALSE
)
}
\arguments{
\item{filename}{The filename of the \code{ZIP} Core Places file.}

\item{dir}{The directory that the file is in.}

\item{exdir}{Name of the directory to unzip to.}

\item{cleanup}{Set to \code{TRUE} to delete all the unzipped files after being read in.}

\item{silent}{Suppress timing messages.}
}
\description{
Feed this function the most recent Core Places file, and it will give you back a \code{data.table} with two columns: \code{safegraph_place_id} and \code{naics_code}. Saving this file is recommended. Then, provide this object to \code{read_shop} or \code{read_many_shop} so that you can use \code{'naics_code'} in the \code{by} argument.
}
\details{
This function is DEPRECATED. It only works with the OLD Core files, and is superseded by the more flexible \code{read_core()} function.
}
\examples{

\dontrun{
# Core-USA-June2020-Release-CORE_POI-2020_05-2020-06-06.zip is a Core places file in the working directory
poi_link <- link_poi_naics('Core-USA-June2020-Release-CORE_POI-2020_05-2020-06-06.zip')
}
}
