% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ma}
\alias{ma}
\title{Seven-Day Moving Average}
\usage{
ma(x, n = 7)
}
\arguments{
\item{x}{The variable to calculate the moving average of.}

\item{n}{The number of lags to cover in the moving average.}
}
\description{
This function returns a (by default) seven-day moving average of the variable passed in. Make sure the data is pre-sorted by date, and grouped by the appropriate grouping. The data should have no gaps in time.
}
\examples{

ma(1:9)

}
