% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naics_2}
\alias{naics_2}
\title{NAICS 4-Digit Code Titles}
\format{
A \code{data.table} with 24 rows and 2 variables:
\describe{
  \item{naics_code}{The NAICS code}
  \item{naics_title}{The title of the NAICS code}
}
}
\source{
\url{US Census NAICS page}
}
\usage{
naics_2
}
\description{
A dataset that links two-digit NAICS codes to their descriptive titles using 2017 NAICS codes.
}
\details{
Notice that some \code{naics_title} values are repeated because they cross several two-digit codes.
}
\keyword{datasets}
