% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naics_4}
\alias{naics_4}
\title{NAICS 4-Digit Code Titles}
\format{
A \code{data.table} with 311 rows and 2 variables:
\describe{
  \item{naics_code}{The NAICS code}
  \item{naics_title}{The title of the NAICS code}
}
}
\source{
\url{US Census NAICS page}
}
\usage{
naics_4
}
\description{
A dataset that links four-digit NAICS codes to their descriptive titles using 2017 NAICS codes.
}
\keyword{datasets}
