% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{norm}
\alias{norm}
\title{Example Normalization Data}
\format{
A \code{data.table} with 7 rows and 7 variables:
\describe{
  \item{date}{The date}
  \item{total_visits}{The total number of visits recorded in SafeGraph on that day}
  \item{total_devices_seen}{The total number of individual devices recorded in SafeGraph on that day}
  \item{total_home_visits}{Total devices with at least one visit to the home location that day}
}
}
\source{
SafeGraph
}
\usage{
norm
}
\description{
The normalization file from the July 1 weekly patterns pull, for the vignette. See [SafeGraph Docs](https://docs.safegraph.com/docs) for full documentation.
}
\keyword{datasets}
