% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{panel}
\alias{panel}
\title{Example Panel Information Data}
\format{
A \code{data.table} with 1155 rows and 4 variables:
\describe{
  \item{start_date}{The first date present in the patterns file}
  \item{state_fips,county_fips}{State and county identifiers}
  \item{number_devices_residing}{The total number of devices with home locations in those counties}
}
}
\source{
SafeGraph
}
\usage{
panel
}
\description{
The home_panel_summary file from the July 1 weekly patterns pull, processed. See [SafeGraph Docs](https://docs.safegraph.com/docs) for full documentation.
}
\keyword{datasets}
