% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_core.R
\name{read_core}
\alias{read_core}
\title{Read SafeGraph Core}
\usage{
read_core(
  filename,
  dir = ".",
  exdir = ".",
  cleanup = TRUE,
  filter = NULL,
  select = NULL,
  silent = FALSE,
  ...
)
}
\arguments{
\item{filename}{The filename of the \code{ZIP} Core Places file.}

\item{dir}{The directory that the file is in.}

\item{exdir}{Name of the directory to unzip to.}

\item{cleanup}{Set to \code{TRUE} to delete all the unzipped files after being read in.}

\item{filter}{A character string describing a logical statement for filtering the data, for example \code{filter = 'naics_code == 512131'} would give you only movie theater POIs. Will be used as an \code{i} argument in a \code{data.table}, see \code{help(data.table)}. Filtering here instead of afterwards can cut down on time and memory demands.}

\item{select}{Character vector of variables to get from the file. Set to \code{NULL} to get all variables. If you plan to link the results to a patterns file, you will probably want to include \code{'safegraph_place_id'} or \code{'placekey'} in this vector. Note that any variables mentioned in \code{filter} MUST be in \code{select} unless \code{select = NULL}.}

\item{silent}{Suppress timing messages.}

\item{...}{Other arguments to be passed to \code{data.table::fread} when reading in the \code{CSV} files inside of the \code{ZIP}. For example, \code{nrows} to only read in a certain number of rows.}
}
\description{
Be aware that the files this is designed to work with are large and this function may take a while to execute. This function takes a single \code{.zip} SafeGraph Core file and reads it in. The output is a \code{data.table}.
}
\details{
If you want \emph{all} of the rows and only want to extract their \code{naics_code}, you can use the deprecated but still-working \code{link_poi_naics()} function.
}
\examples{

\dontrun{
# Location of our CORE file
dir <- '../SafeGraph/2020/10/'
filename <- 'Core-USA-Oct-CORE_POI-2020_09-2020-10-19.zip'

# Let's only get retail POIs in California
# And
locations <- read_core(filename = filename,
                       dir = dir,
                       filter = 'region == "CA" & floor(naics_code/10000) \%in\% 44:45')
}
}
