% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_many_files.R
\name{read_many_csvs}
\alias{read_many_csvs}
\title{Read and row-bind many CSVs}
\usage{
read_many_csvs(
  dir = ".",
  recursive = FALSE,
  filelist = NULL,
  makedate = FALSE,
  ...
)
}
\arguments{
\item{dir}{Name of the directory the files are in.}

\item{recursive}{Search in all subdirectories as well.}

\item{filelist}{Optionally specify only a subset of the filename to read in (can contain paths).}

\item{makedate}{Use \code{year}, \code{month}, and \code{day} columns in the data to create a \code{date} variable. Works with normalization files.}

\item{...}{Other arguments to pass to \code{data.table::fread}.}
}
\description{
This accepts a directory. It will load every \code{csv} or \code{csv.gz} in that folder and attempt to row-bind them together. You can alternately specify a list of files if you don't want everything in the folder. This is designed for use with the normalization and home-summary files as downloaded from AWS.
}
\examples{
\dontrun{

# The current working directory contains all the normalization .csv files
normalization <- read_many_csvs(makedate = TRUE)

}
}
