% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_monthly_shop.R
\name{read_monthly_shop}
\alias{read_monthly_shop}
\title{Read Monthly Patterns from the SafeGraph Shop}
\usage{
read_monthly_shop(filename, dir = "./")
}
\arguments{
\item{filename}{The filename of the ZIP file from the SafeGraph shop which contains patterns, normalization, etc.}

\item{dir}{The directory in which the ZIP file sits.}
}
\description{
This function takes a monthly patterns file as downloaded from the SafeGraph shop and returns \code{visits_by_day} on a daily level in three formats: by NAICS code by state, by brand by state, and by \code{sub_category} by state. Plus, a normalization file.
}
