% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sg_aws.R
\name{read_sg_aws}
\alias{read_sg_aws}
\title{Read Weekly Patterns as Downloaded from AWS}
\usage{
read_sg_aws(
  filename,
  dir = ".",
  by = NULL,
  expand_int = NULL,
  expand_cat = NULL,
  expand_name = NULL,
  multi = NULL,
  select = NULL,
  gen_fips = TRUE,
  start_date = NULL
)
}
\arguments{
\item{filename}{The filename of the \code{.csv.gz} file.}

\item{dir}{The directory in which the file sits.}

\item{by}{A character vector giving the variable names of the level to be collapsed to using \code{sum(na.rm=TRUE)}. The resulting data will have X rows per unique combination of \code{by}, where X is 1 if no expand variables are specified, or the length of the expand variable if specified. Set to \code{NULL} to aggregate across all initial rows, or set to \code{FALSE} to not aggregate at all (this will also add an \code{initial_rowno} column showing the original row number).}

\item{expand_int}{A character variable with the name of The first e JSON variable in integer format ([1,2,3,...]) to be expanded into rows. Cannot be specified along with \code{expand_cat}.}

\item{expand_cat}{A JSON variable in categorical format ({A: 2, B: 3, etc.}) to be expanded into rows.  Ignored if \code{expand_int} is specified.}

\item{expand_name}{The name of the new variable to be created with the category index for the expanded variable.}

\item{multi}{A list of named lists, for the purposes of creating a list of multiple processed files. This will vastly speed up processing over doing each of them one at a time. Each named list has the entry \code{name} as well as any of the options \code{by, expand_int, expand_cat, expand_name} as specified above. If specified, will override other entries of \code{by}, etc..}

\item{select}{Character vector of variables to get from the file. Set to \code{NULL} to get all variables.}

\item{gen_fips}{Set to \code{TRUE} to use the \code{poi_cbg} variable to generate \code{state_fips} and \code{county_fips} variables. This will also result in \code{poi_cbg} being converted to character.}

\item{start_date}{The first date in the file, as a date object. If omitted, will assume that the filename begins YYYY-MM-DD.}

\item{...}{Other arguments to be passed to \code{data.table::fread} when reading in the file. For example, \code{nrows} to only read in a certain number of rows.}
}
\description{
Be aware that the files this is designed to work with are large and this function may take a while to execute. This function takes a single \code{.csv.gz} patterns file as provided for data via AWS and reads it in. The output is a \code{data.table} (or a list of them if multiple are specified) including the file \code{filename} collapsed and expanded in different ways.
}
