% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_shop.R
\name{read_shop}
\alias{read_shop}
\title{Read a ZIP file with patterns and other data as it comes from the SafeGraph Shop}
\usage{
read_shop(
  file,
  dir = ".",
  keeplist = c("patterns", "normalization_stats.csv", "home_panel_summary.csv",
    "visit_panel_summary.csv", "brand_info.csv"),
  exdir = dir,
  cleanup = TRUE,
  by = NULL,
  fun = sum,
  na.rm = TRUE,
  filter = NULL,
  expand_int = NULL,
  expand_cat = NULL,
  expand_name = NULL,
  multi = NULL,
  select = NULL,
  gen_fips = TRUE,
  start_date = NULL,
  ...
)
}
\arguments{
\item{file}{The filename of the \code{.zip} file from the shop.}

\item{dir}{The directory the file is in.}

\item{keeplist}{Character vector of the files in the ZIP to read in. Use \code{'patterns'} to refer to the patterns files.}

\item{exdir}{Name of the directory to unzip to.}

\item{cleanup}{Set to \code{TRUE} to delete all the unzipped files after being read in.}

\item{by, fun, na.rm, filter, expand_int, expand_cat, expand_name, multi, select, gen_fips, ...}{Other arguments to be passed to \code{read_patterns}, specified as in \code{help(read_patterns)}.}

\item{start_date}{An argument to be passed to \code{read_patterns} giving the first date present in the file, as a date object. When using \code{read_shop} this should usually be included, since the patterns file names in the shop files are not in a format \code{read_patterns} can pick up on automatically.}
}
\description{
This will open up a ZIP file from the SafeGraph shop and will read all of the data in, performing processing of the patterns files using \code{read_patterns}.
}
\details{
The result will be a named list with each of the components of the data.
}
