% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_weekly_aws.R
\name{read_weekly_aws}
\alias{read_weekly_aws}
\title{Read Weekly Patterns as Downloaded from AWS}
\usage{
read_weekly_aws(filename, dir)
}
\arguments{
\item{filename}{The filename of the \code{.csv.gz} file.}

\item{dir}{The directory in which the file sits.}
}
\description{
This function takes a single \code{.csv.gz} patterns file as provided for weekly data via AWS and reads in the \code{visits_by_day} information. The result is two data sets, both by-state-by-day, one also by-brand, and one by-NAICS.
}
