% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_to_date.R
\name{scale_to_date}
\alias{scale_to_date}
\title{Scale data relative to its value on a date}
\usage{
scale_to_date(
  data,
  date,
  date_var = "date",
  adj_vars,
  by = NULL,
  growth = FALSE,
  format_percent = FALSE,
  accuracy = 0.1
)
}
\arguments{
\item{data}{A \code{data.frame}, \code{tibble}, or \code{data.table}. Note that a \code{data.table} will be returned.}

\item{date}{The date you'd like everything relative to, as a date object.}

\item{date_var}{The name of the date variable.}

\item{adj_vars}{Character vector of the variable names you'd like adjusted to be relative to the date.}

\item{by}{Character vector of the variable names you'd like the operation to be performed by. There should only be one observation for which \code{date_var == date} within each combination of the \code{by} variables, or else your results will be arbitrary.}

\item{growth}{Set to \code{TRUE} to get \code{new/old - 1} (i.e. a percentage growth). Set to \code{FALSE} to get \code{new/old} (i.e. a relative value).}

\item{format_percent}{Set to \code{TRUE} to get back a formatted percentage, i.e. "50%", instead of a number.}

\item{accuracy}{If \code{format_percent = TRUE}, the number of digits after the decimal place to round to, as in \code{scales::percent}.}
}
\description{
Pick a date and provide some variables. Those variables will be adjusted to be relative to their value on that date. Usually used to calculate foot traffic growth relative to a certain date.
}
