% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_to_date.R
\name{scale_yoy}
\alias{scale_yoy}
\title{Calculate year-on-year change}
\usage{
scale_yoy(
  data,
  yoy_vars,
  date_var = "date",
  leap_year_fillin = TRUE,
  by = NULL,
  growth = TRUE,
  format_percent = FALSE,
  accuracy = 0.1
)
}
\arguments{
\item{data}{Any type of data set that can be coerced to a \code{data.table}.}

\item{yoy_vars}{String vector of the variable names you want to calculate year-on-year change for.}

\item{date_var}{The name of the date variable, as a string. Must be formatted as Date objects.}

\item{leap_year_fillin}{If the date is Feb. 29, the previous year will not have a Feb. 29. Set to \code{TRUE} to fill in by linear interpolation. If set to \code{TRUE}, returned data will be sorted by \code{by} and \code{date_var}.}

\item{by}{Character vector of the variable names you'd like the operation to be performed by. There should only be one observation per date per combination of \code{by}.}

\item{growth}{Set to \code{TRUE} to get \code{new/old - 1} (i.e. a percentage growth). Set to \code{FALSE} to get \code{new/old} (i.e. a relative value).}

\item{format_percent}{Set to \code{TRUE} to get back a formatted percentage, i.e. "50\%", instead of a number.}

\item{accuracy}{If \code{format_percent = TRUE}, the number of digits after the decimal place to round to, as in \code{scales::percent}.}
}
\description{
This takes a data set with a date variable and calculates year-on-year changes for a set of variables of your choice. Returns a \code{data.table}.
}
\details{
This will add new variables using \code{yoy_vars}, adding \code{lag} and \code{YOY} variants.
}
