% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_info}
\alias{state_info}
\title{State Information}
\format{
A \code{data.table} with 51 rows and 4 variables:
\describe{
  \item{statename}{The full name of the state / province}
  \item{CensusRegion}{The broad Census regions}
  \item{region}{The state's two-digit abbreviation / the province's international alpha code}
  \item{state_fips}{State FIPS code / Canadian SGC code}
  \item{iso_country_code}{Indicator for US or Canada}
}
}
\source{
\url{US Census}
}
\usage{
state_info
}
\description{
A dataset that links state (and Washington DC) names, FIPs codes, two-letter abbreviations (called "region" because this is what it is called in SafeGraph files that use it), and Census regions. Can be merged with \code{fips_to_names} using \code{state_fips} and \code{statename}.
}
\details{
This also includes Canadian data on provinces.

Note that this is a data set purely of Canadian provinces, US *states*, and DC. Some SafeGraph files contain information on \code{region} values of \code{GU} (Guam), \code{PR} (Puerto Rico), etc., but those will be lost if merging with \code{state_info}.
}
\keyword{datasets}
