% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_unzip.R
\name{temp_unzip}
\alias{temp_unzip}
\title{Unzip a Data File and Read It}
\usage{
temp_unzip(filename, fun = data.table::fread, ...)
}
\arguments{
\item{filename}{The filename of the zip file to be unzipped.}

\item{fun}{The function to use to read in the resulting unzipped data file.}

\item{...}{Additional arguments to be passed to \code{fun}. Often, this is \code{select} to only read in some of the variables.}
}
\description{
SafeGraph patterns files come in \code{.csv.gz} format. \code{temp_unzip} will unzip the file, read it in using \code{fun} (usually \code{fread}), and then delete the unzipped \code{.csv} file.
}
\details{
I stole this function from a StackExchange answer many years ago and have copied it so many times I no longer know where it came from. Every day we offer thanks to the author.
}
