% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbg_pop}
\alias{cbg_pop}
\title{Unweighted Population by Census Block Group}
\format{
A \code{data.table} with 220333 rows and 2 variables:
\describe{
  \item{poi_cbg}{Census Block Group ID, named for easy merging with the patterns data.}
  \item{unweighted_pop}{Population from the 2016 American Community Survey (the "unweighted" part is outdated but kept for consistency with old code).}
}
}
\source{
\url{https://docs.safegraph.com/docs/open-census-data}
}
\usage{
cbg_pop
}
\description{
A dataset containing the unweighted population by Census Block Group from the Open Census file. Use with \code{fips_from_cbg} to get state and county FIPS codes from the CBG ID.
}
\keyword{datasets}
