% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fips_from_cbg}
\alias{fips_from_cbg}
\title{Pull state and county FIPS (US) or province and census division (CA) from CBG code}
\usage{
fips_from_cbg(cbg, return = "both")
}
\arguments{
\item{cbg}{CBG code, in numeric or string form. To aid speed since this function is called millions of times, \code{cbg} is not checked to ensure it is a valid CBG identifier.}

\item{return}{Set to 'state' to get back only state FIPS, 'county' for only county, or 'both' for a list of both (state then county).}
}
\description{
This function takes a CBG code (as numeric or string) and returns the state and county FIPS codes associated with it.
}
\details{
The syntax for this function was developed before the Canadian data was introduced, so it is definitely US-first, down to the function name, with Canadian additions tacked on. Sorry neighbors to the North. Canadian province ("state") and census division ("county") identifiers will be preceded with \code{"CA:"} as in the SafeGraph \code{cbg} variable.

This function now returns character values rather than numeric, to account for the Canadian data.

Why does this produce a list and not a vector? For \code{data.table} usage.
}
\examples{

a_cbg <- '560610112022'
fips_from_cbg(a_cbg)

# Use with data.table!
DT <- data.table::data.table(cbg = c('560610112022','10310112022'))
DT[,c('state_fips','county_fips') := fips_from_cbg(cbg)]

}
