% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rbind_by_list_pos}
\alias{rbind_by_list_pos}
\title{Row-binds data.tables in a list of lists}
\usage{
rbind_by_list_pos(dtl, ignore_names = FALSE)
}
\arguments{
\item{dtl}{List of lists of \code{data.table}s.}

\item{ignore_names}{If the list is named, match objects across lists only by their position in the list and not by their names.}
}
\description{
This function takes a list of lists of \code{data.table}s, and then row-binds them by position or name. For example, if passed \code{list(list(first=A,second=B),list(first=C,second=D))}, you would get back \code{list(first=rbind(A,C),second=rbind(B,D))}.
}
