% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_distancing.R
\name{read_distancing}
\alias{read_distancing}
\title{Read in Stay-at-Home Data}
\usage{
read_distancing(start, end, dir)
}
\arguments{
\item{start}{Date object with the starting date to read in stay-at-home data.}

\item{end}{Ending date to read stay-at-home data to.}

\item{dir}{The folder in which the "2020" (etc.) folder resides.}
}
\description{
Takes a folder of stay-at-home Safegraph data structured how it comes from AWS (i.e. folders 2020/04/03 for April 3 2020) and reads them in.
}
