% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_many_files.R
\name{read_many_patterns}
\alias{read_many_patterns}
\title{Read and row-bind many patterns files}
\usage{
read_many_patterns(
  dir = ".",
  filelist = NULL,
  by = NULL,
  fun = sum,
  na.rm = TRUE,
  filter = NULL,
  expand_int = NULL,
  expand_cat = NULL,
  expand_name = NULL,
  multi = NULL,
  naics_link = NULL,
  select = NULL,
  gen_fips = TRUE,
  start_date = NULL,
  silent = FALSE,
  ...
)
}
\arguments{
\item{dir}{Name of the directory the files are in.}

\item{filelist}{Optionally specify only a subset of the files to read in.}

\item{by, fun, na.rm, filter, expand_int, expand_cat, expand_name, multi, naics_link, select, gen_fips, silent, ...}{Arguments to be passed to \code{read_patterns}, specified as in \code{help(read_patterns)}.}

\item{start_date}{A vector of dates giving the first date present in each zip file, to be passed to \code{read_patterns} giving the first date present in the file, as a date object.}
}
\description{
This accepts a directory. It will use \code{read_patterns} to load every \code{csv.gz} in that folder, assuming they are all patterns files. It will then row-bind together each of the produced processed files.
}
\details{
Note that after reading in data, if \code{gen_fips = TRUE}, state and county names can be merged in using \code{data(fips_to_names)}.
}
