% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_many_files.R
\name{read_many_patterns}
\alias{read_many_patterns}
\title{Read and row-bind many patterns files}
\usage{
read_many_patterns(
  filelist,
  dir = ".",
  by = NULL,
  fun = sum,
  na.rm = TRUE,
  expand_int = NULL,
  expand_cat = NULL,
  expand_name = NULL,
  multi = NULL,
  select = NULL,
  gen_fips = TRUE,
  start_date = NULL,
  ...
)
}
\arguments{
\item{filelist}{Optionally specify only a subset of the filename to read in.}

\item{dir}{Name of the directory the files are in.}

\item{by, fun, na.rm, expand_int, expand_cat, expand_name, multi, select, gen_fips, start_date, ...}{Arguments to be passed to \code{read_patterns}, specified as in \code{help(read_patterns)}.}
}
\description{
This accepts a directory. It will use \code{read_patterns} to load every \code{csv.gz} in that folder, assuming they are all patterns files. It will then row-bind together each of the produced processed files.
}
