% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safegraph_aws.R
\name{safegraph_aws}
\alias{safegraph_aws}
\title{Download SafeGraph data from AWS COVID Response}
\usage{
safegraph_aws(
  path = ".",
  dataset,
  bucket_only = FALSE,
  base_url = "s3.wasabisys.com",
  key,
  secret,
  region = "",
  prefix = "",
  prefix_is_dir = FALSE,
  max_print = 1,
  ...
)
}
\arguments{
\item{path}{The local directory to synchronize.}

\item{dataset}{The SafeGraph bucket to get from. Can be "weekly" (OLD VERSION), "weekly-new" (new method since December 2020), "weekly-backfill" (the new method for times before December 2020), "monthly" (OLD VERSION, but also includes the backfill data for the new version), "monthly-backfill" (method since Dec 2020), "distancing", "core" (before Nov 2020), "core-new" (Nov 2020 or later), "geo-supplement", "neighborhood" or, to get the baseline bucket, "none". v2 versions always selected.}

\item{bucket_only}{Instead of doing an \code{aws.s3::s3sync} call, just return the correct bucket as a string. Then you can use that to do your own \code{aws.s3::s3sync} call, or work with the AWS CLI.}

\item{base_url}{The base URL to pull the data from.}

\item{key}{A character string containing an AWS Access Key ID.}

\item{secret}{A character string containing an AWS Secret Access Key.}

\item{region}{A character string containing the AWS region.}

\item{prefix}{Leading part of the objects in the bucket must have this prefix. For example, to download social distancing data only from 2020, set this to "2020/".}

\item{prefix_is_dir}{If \code{FALSE}, the files matching \code{prefix} will be downloaded directly to \code{path}, which may not be desired behavior if \code{prefix} contains a directory (you probably want the directory structure to match!). Set to \code{TRUE} to, in effect, replace \code{path} with \code{paste0(path, prefix)} and so download files to the appropriate folder. Don't use if \code{prefix} also contains file characteristics like extension. This is \code{prefix_IS_dir}, not \code{prefix_CONTAINS_dir}.}

\item{max_print}{Temporarily set \code{options(max.print)} to this value. This will massively speed up the function, as \code{aws.s3::s3sync} likes to print the full list of files on the server before moving on. The option will be returned to its original value afterwards. Set to \code{NULL} to not alter any options.}

\item{...}{Additional parameters to be sent to \code{aws.s3::s3sync} and from there on to \code{aws.s3:s3HTTP}. "direction" will be ignored.}
}
\description{
This is a thin wrapper for \code{aws.s3::s3sync} that will aim you at the right directory to synchronize.
}
\details{
This function doesn't add too much, but it does make the default behavior you probably want a bit easier. If you plan to specify the \code{aws.s3::s3sync} "bucket" option yourself, this function is largely useless.

See catalog.safegraph.io for more description of the various buckets.
}
\examples{

\dontrun{

# Download all the recent weekly-patterns files to the working directory
safegraph_aws(dataset = 'weekly-new', key = 'MYINFO', secret = 'MYOTHERINFO')

}

}
